+++ src/index.ts
@@ -10,7 +10,7 @@ export * from "./classes/FritterResponse.js";
 export * as BodyParserMiddleware from "./middlewares/body-parser.js";
 export * as CorsMiddleware from "./middlewares/CorsMiddleware.js";
 export * as CurrentPageNumberMiddleware from "./middlewares/CurrentPageNumberMiddleware.js";
-export * as ForceSslMiddleware from "./middlewares/force-ssl.js";
+export * as ForceSslMiddleware from "./middlewares/ForceSslMiddleware.js";
 export * as LogRequestMiddleware from "./middlewares/log-request.js";
 export * as RouterMiddleware from "./middlewares/router.js";
 export * as StaticMiddleware from "./middlewares/static.js";